feat(plugin-source-build): support analyze exports field
