+++ e2e/cases/source-build/utils/package.json
@@ -2,7 +2,6 @@
   "name": "@e2e/source-build-utils",
   "private": true,
   "version": "1.0.0",
-  "source": "./src/index.ts",
   "types": "./dist/types/index.d.ts",
   "exports": {
     ".": {
