+++ packages/monorepo-utils/src/project-utils/index.ts
@@ -1,2 +1,3 @@
 export * from './getDependentProjects';
+export * from './getExportsSourceDirs';
 export * from './filter';
