+++ packages/monorepo-utils/src/project/project.ts
@@ -1,7 +1,8 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import type { INodePackageJson, ExportsConfig } from '../types/packageJson';
+import type { INodePackageJson } from '../types/packageJson';
 import { readPackageJson } from '../utils';
+import { getExportsSourceDirs } from '../project-utils';
 import { PACKAGE_JSON } from '../constants';
 
 export class Project {
@@ -95,12 +96,10 @@ export class Project {
     const pkgJson = this.getMetaData() as INodePackageJson &
       Record<string, string>;
 
-    if (!(sourceField in pkgJson)) {
-      throw new Error(`${this.name} 的 package.json 没有 ${sourceField} 字段`);
-    }
-    const sourceDir = path.normalize(pkgJson[sourceField]);
     // normalize strings
-    const sourceDirs = [sourceDir];
+    const sourceDirs = pkgJson[sourceField]
+      ? [path.normalize(pkgJson[sourceField])]
+      : [];
 
     if (checkExports) {
       /**
@@ -111,32 +110,18 @@ export class Project {
        *   }
        * },
        */
-      const exportsSourceDirs = this.#getExportsSourceDirs(
+      const exportsSourceDirs = getExportsSourceDirs(
         pkgJson.exports ?? {},
         sourceField,
       );
       sourceDirs.push(...exportsSourceDirs);
     }
 
-    return this.#getCommonRootPaths(sourceDirs);
-  }
-
-  #getExportsSourceDirs(exportsConfig: ExportsConfig, sourceField: string) {
-    const exportsSourceDirs: string[] = [];
-
-    for (const moduleRules of Object.values(exportsConfig)) {
-      if (
-        typeof moduleRules === 'object' &&
-        typeof moduleRules[sourceField] === 'string'
-      ) {
-        exportsSourceDirs.push(
-          path.normalize(moduleRules[sourceField] as string),
-        );
-      }
+    if (!sourceDirs.length) {
+      throw new Error(`${this.name} 的 package.json 没有 ${sourceField} 字段`);
     }
 
-    // normalize strings
-    return exportsSourceDirs;
+    return this.#getCommonRootPaths(sourceDirs);
   }
 
   /**
